requirements.txt
pyyaml>=6.0
openai>=1.0.0
torch>=2.0.0
tqdm>=4.65.0
anthropic>=0.5.0
vllm>=0.2.0
nltk>=3.8.1
google-genai>=1.0.0

